corrección de errores en el controlador e implementacion del paquete de seguridad y login con JWT

+++ apiRest/src/main/java/com/portfolio/api/Repository/EducationRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface EducationRepository extends JpaRepository<Education, Long> {
+public interface EducationRepository extends JpaRepository<Education, Integer> {
     
 }
