+++ apiRest/src/main/java/com/portfolio/api/Repository/ExperienceRepository.java
@@ -1,4 +1,3 @@
-
 package com.portfolio.api.Repository;
 
 import com.portfolio.api.Model.Experience;
@@ -6,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExperienceRepository extends JpaRepository<Experience, Long> {
-    
+public interface ExperienceRepository extends JpaRepository<Experience, Integer> {
 }
