+++ apiRest/src/main/java/com/portfolio/api/Repository/PersonaRepository.java
@@ -1,4 +1,3 @@
-
 package com.portfolio.api.Repository;
 
 import com.portfolio.api.Model.Persona;
@@ -6,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonaRepository extends JpaRepository<Persona, Long> {
-    
+public interface PersonaRepository extends JpaRepository<Persona, Integer> {
 }
